Filter clothes by category and brand
The user sees a list of clothing. There is a filter button to show clothes by a selected category and/or a brand. There's also a reset filter button that clears the filter and shows all the garments available.

The app takes in the data from an array of objects. The app is built by sma
ll specific functions for each action such as:

– show(): shows the clothing list for the user.
– clearList(): removes all the items in the window. This function is used inside the show() function to avoid showing a duplicate list.
– filter(): filters the array based on the selected category or brand. It also works by filtering by both category and brand.
– resetFilter(): shows the original complete list of garments and resets the radio buttons.