Check if user is human with a checkbox
To be able to filter, the user has to check the “I’m not a robot” checkbox, otherwise it will appear a message asking to do it. Once the user clicks the checkbox, the message is removed.

The new function checkHuman() manages the three different cases after the user clicks the filter button:
1. If it’s checked, continue filtering.
2. If it’s not checked, show message.
3. If it’s still not checked and there’s already a message, stop filtering.

The reset button now clears the checkbox as well.

Fixed index.html so that label and input id match.